+++ libs/ml/examples/word2vec_SGNS_dataloader/main.cpp
@@ -26,7 +26,7 @@
 #include "ml/ops/loss_functions.hpp"
 #include "ml/optimisation/adam_optimiser.hpp"
 #include "ml/optimisation/sgd_optimiser.hpp"
-#include "ml/utilities/model_saver.hpp"
+#include "ml/utilities/graph_saver.hpp"
 #include "word2vec_utilities.hpp"
 
 #include <iostream>
