+++ libs/ml/examples/word2vec_graph_evaluation/main.cpp
@@ -20,7 +20,7 @@
 #include "ml/core/graph.hpp"
 #include "ml/dataloaders/word2vec_loaders/sgns_w2v_dataloader.hpp"
 #include "ml/layers/skip_gram.hpp"
-#include "ml/utilities/model_saver.hpp"
+#include "ml/utilities/graph_saver.hpp"
 #include "word2vec_utilities.hpp"
 
 #include <stdexcept>
